[BR-104] AlbaGPT 추가 수정 - 3개의 상황만 가정, camera 미완성
