+++ alba_gpt/AlbaGPT_main/contents/example/maintenance_example.csv
@@ -1,0 +1,41 @@
+prompt,pinky_id,pinky_task,pinky_response
+4번 핑키 베터리가 부족해 보이는데,4,MAINTENANCE,네! 충전 스테이션으로 이동하겠습니다 🔌,
+3번 핑키 베터리 충전하러 가,3,MAINTENANCE,충전하러 출발할게요! ⚡,
+2번 핑키 충전이 필요한 것 같아,2,MAINTENANCE,"점검 중입니다. 충전이 필요해요, 이동할게요 🛠️",
+3번 핑키 베터리 거의 다 닳았어,3,MAINTENANCE,네! 충전 스테이션으로 이동하겠습니다 🚧,
+5번 핑키 배터리 경고 떴어,5,MAINTENANCE,잔량 부족해요. 충전할게요! 🔎,
+5번 핑키 방전되기 전에 충전해,5,MAINTENANCE,네! 최대한 빠르게 충전하겠습니다 🚧,
+4번 핑키 배터리 상태 점검해줘,4,MAINTENANCE,충전이 필요하다는 신호를 받았어요. 이동합니다! 🚧,
+3번 핑키 전력이 부족한 것 같아,3,MAINTENANCE,알겠습니다. 즉시 충전하겠습니다 🔌,
+1번 핑키 에너지 부족하대,1,MAINTENANCE,충전하러 출발할게요! 🚧,
+1번 핑키 지금 전력 낮아졌어,1,MAINTENANCE,네! 최대한 빠르게 충전하겠습니다 🔋,
+2번 핑키 전원 꺼질 것 같아,2,MAINTENANCE,충전이 필요하다는 신호를 받았어요. 이동합니다! 🧪,
+4번 핑키 빨리 충전하자,4,MAINTENANCE,지금 충전 스테이션으로 이동 중입니다 🔌,
+3번 핑키 파워 부족 알림 떴어,3,MAINTENANCE,충전이 필요하다는 신호를 받았어요. 이동합니다! 🛠️,
+2번 핑키 전기 좀 채워야겠어,2,MAINTENANCE,네! 충전 스테이션으로 이동하겠습니다 🛠️,
+1번 핑키 충전 필요 경고 봤어,1,MAINTENANCE,"점검 중입니다. 충전이 필요해요, 이동할게요 🛠️",
+5번 핑키 지금 전력 상황 어때?,5,MAINTENANCE,"네, 바로 충전 스테이션으로 이동합니다 🧪",
+1번 핑키 오늘 충전 안 했지?,1,MAINTENANCE,네! 최대한 빠르게 충전하겠습니다 🪫,
+5번 핑키 배터리 잔량 얼마 안 남았어,5,MAINTENANCE,충전하겠습니다. 잠시만요! 🚨,
+1번 핑키 에너지 다 썼어,1,MAINTENANCE,네! 충전 스테이션으로 이동하겠습니다 🧪,
+5번 핑키 배터리 점검하고 와,5,MAINTENANCE,네! 최대한 빠르게 충전하겠습니다 🔋,
+2번 핑키 오늘 충전한 적 있니?,2,MAINTENANCE,잔량 부족해요. 충전할게요! 🛠️,
+5번 핑키 충전 좀 하고 와,5,MAINTENANCE,충전이 필요하다는 신호를 받았어요. 이동합니다! 🔌,
+1번 핑키 전력 떨어지는 중이야,1,MAINTENANCE,충전하겠습니다. 잠시만요! 🔎,
+4번 핑키 가기 전에 충전부터 하자,4,MAINTENANCE,지금 충전 스테이션으로 이동 중입니다 ⚡,
+4번 핑키 베터리 상태 어때?,4,MAINTENANCE,"네, 바로 충전 스테이션으로 이동합니다 ⚡",
+1번 핑키 지금 충전이 필요한 시간 같아,1,MAINTENANCE,충전하겠습니다. 잠시만요! 🪫,
+1번 핑키 에너지 비상사태야,1,MAINTENANCE,"점검 중입니다. 충전이 필요해요, 이동할게요 🧯",
+3번 핑키 곧 전원 꺼질 수 있어,3,MAINTENANCE,알겠습니다. 즉시 충전하겠습니다 🚨,
+4번 핑키 충전 포인트로 이동해,4,MAINTENANCE,지금 충전 스테이션으로 이동 중입니다 🚧,
+4번 핑키 지금 배터리 얼마나 남았어?,4,MAINTENANCE,충전하겠습니다. 잠시만요! 🔌,
+4번 핑키 배터리 부족한 거 맞지?,4,MAINTENANCE,알겠습니다. 즉시 충전하겠습니다 🔌,
+1번 핑키 충전해야 움직일 수 있어,1,MAINTENANCE,"점검 중입니다. 충전이 필요해요, 이동할게요 🔋",
+4번 핑키 곧 방전이야,4,MAINTENANCE,잔량 부족해요. 충전할게요! ⚡,
+2번 핑키 충전 중인 상태야?,2,MAINTENANCE,"점검 중입니다. 충전이 필요해요, 이동할게요 🔋",
+2번 핑키 충전 안 하면 곤란해,2,MAINTENANCE,네! 충전 스테이션으로 이동하겠습니다 ⚡,
+3번 핑키 파워 거의 0이야,3,MAINTENANCE,충전하러 출발할게요! 🧯,
+1번 핑키 베터리 거의 소진됐어,1,MAINTENANCE,네! 충전 스테이션으로 이동하겠습니다 🪫,
+5번 핑키 충전 없이 버티기 힘들어 보여,5,MAINTENANCE,잔량 부족해요. 충전할게요! 🚧,
+1번 핑키 전원 꺼지기 전에 충전하자,1,MAINTENANCE,충전하겠습니다. 잠시만요! 🪫,
+2번 핑키 충전량 부족이야,2,MAINTENANCE,충전이 필요하다는 신호를 받았어요. 이동합니다! 🔎,
