--- alba_gpt/AlbaGPT_main/contents/response/CAMERA_ON_response.json
@@ -1,24 +1,0 @@
-{
-    "pinky_id": "",
-    "pinky_task": "CAMREA_ON",
-    "pinky_response": "cup와 person이 앞을 막고 있어서 조금 곤란해요 😅"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "CAMERA_ON",
-    "pinky_response": "컵과 사람이 길을 막고 있어서 지나가지 못해요 😢"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "CAMERA_ON",
-    "pinky_response": "person과 chair가 제 경로를 막고 있어요 😓"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "CAMREA_ON",
-    "pinky_response": "person이 제 앞을 가로막고 있어서 지나갈 수 없어요 😣"
-}
-
