--- alba_gpt/AlbaGPT_main/contents/response/CLEANING_response.json
@@ -1,182 +1,0 @@
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블 청소를 시작하겠습니다 🧼",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "4번 테이블을 깨끗하게 청소하겠습니다 🧼",
-    "table_id": "4번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블 청소를 시작하겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "4번 테이블이 더럽습니다! 즉시 청소하겠습니다 🧹",
-    "table_id": "4번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗이 청소하겠습니다 🧼",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗하게 청소하러 가겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "4번 테이블이 더럽습니다. 즉시 청소하겠습니다 🧹",
-    "table_id": "4번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "네! 3번 테이블을 깨끗하게 청소하러 가겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "네! 4번 테이블을 깨끗하게 청소하러 가겠습니다 🧹",
-    "table_id": "4번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗하게 청소하러 가겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "4번 테이블이 더럽습니다. 즉시 청소하겠습니다 🧹",
-    "table_id": "4번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗하게 청소하러 가겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "4번 테이블을 깨끗하게 청소하러 가겠습니다 🧹",
-    "table_id": "4번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗하게 청소하겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗하게 청소하겠습니다 🧼",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "네! 3번 테이블을 깨끗하게 청소하러 가겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗하게 청소하겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗하게 청소하겠습니다 🧹",
-    "table_id": "3"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "네! 3번 테이블을 깨끗하게 정리하러 가겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "네! 4번 테이블을 깨끗하게 정리하러 가겠습니다 🧹",
-    "table_id": "4번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "네! 3번 테이블을 깨끗하게 정리하러 가겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "네! 4번 테이블을 깨끗하게 정리하러 가겠습니다 🧹",
-    "table_id": "4번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗하게 청소하겠습니다 🧼",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "4번 테이블을 깨끗하게 청소하겠습니다 🧼",
-    "table_id": "4번 테이블"
-}
-
-{
-    "pinky_id": "4",
-    "pinky_task": "CLEANING",
-    "pinky_response": "3번 테이블을 깨끗하게 청소하겠습니다 🧹",
-    "table_id": "3번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "CLEANING",
-    "pinky_response": "4번 테이블이 더럽습니다! 즉시 청소하겠습니다 🧹",
-    "table_id": "4번 테이블"
-}
-
