--- alba_gpt/AlbaGPT_main/contents/response/EMERGENCY_response.json
@@ -1,91 +1,0 @@
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "지금 매장에 도둑이 들었습니다! 즉시 대피하세요! 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "핑키들, 지금 매장에 도둑이 들었어! 즉시 대피 안내하겠습니다 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "핑키들, 지금 매장에 도둑이 들었어! 즉시 대처하겠습니다! 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "핑키들, 지금 매장에 도둑이 들었어! 즉시 경계 태세를 갖추고 손님들을 안전하게 대피시키겠습니다!",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "핑키들, 지금 매장에 도둑이 들었어! 즉시 경계를 강화하고 손님들을 안전하게 대피시켜야 해!",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "핑키들, 지금 매장에 도둑이 들었어! 즉시 경계 태세를 갖추고 손님들을 안전하게 대피시켜야 합니다! 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "핑키들, 지금 매장에 도둑이 들었어! 즉시 경고 방송하고 대피 안내하겠습니다 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "매장에 도둑이 들었습니다! 즉시 경고하고 대피를 안내하겠습니다! 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "매장에 도둑이 들었습니다! 즉시 경고 방송을 시작하겠습니다! 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "핑키들, 지금 매장에 도둑이 들었어! 즉시 대처하겠습니다!",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "도둑이 들었습니다! 즉시 경고하고 대피를 안내하겠습니다!",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "지금 매장에 도둑이 들었어요! 즉시 대피하세요! 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "도둑이 들어왔습니다! 즉시 경고하고 대피를 안내하겠습니다! 🚨",
-    "table_id": ""
-}
-
