+++ alba_gpt/AlbaGPT_main/contents/response/GREETINGS_response.json
@@ -1,28 +1,4 @@
 {
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎉"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루가 시작됐어요 🎉"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루가 시작됐어요 🎉"
-}
-
-{
     "pinky_id": "",
     "pinky_task": "GREETINGS",
     "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
@@ -32,131 +8,5 @@
     "pinky_id": "1",
     "pinky_task": "GREETINGS",
     "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "사장님, 오늘 기분이 좋으신가요? 저도 기대됩니다! 😊"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "사장님! 오늘 기분이 어떠신가요? 😊"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루가 될 것 같아요 🎉"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎉"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎉"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎉"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 좋습니다 사장님! 신나는 하루가 기대돼요 🎉"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루가 될 것 같아요 🎉"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "사장님, 오늘 기분이 좋으신가요? 저도 기대됩니다! 😊"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎉"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋아요 사장님! 함께 즐거운 하루 만들어봐요 😊"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 함께 즐거운 하루 만들어봐요 😊"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루가 될 것 같아요 🎉"
 }
