--- alba_gpt/AlbaGPT_main/contents/response/IDLE_response.json
@@ -1,28 +1,0 @@
-{
-    "pinky_id": "2",
-    "pinky_task": "IDLE",
-    "pinky_response": "고생했어요! 원래 하던 일로 복귀합니다 😊",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "IDLE",
-    "pinky_response": "고생했어요! 원래 하던 일로 복귀합니다 😊",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "IDLE",
-    "pinky_response": "네 원래 하던 일로 복귀합니다! 😊",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "IDLE",
-    "pinky_response": "네! 원래 하던 일로 복귀합니다! 😊",
-    "table_id": ""
-}
-
