--- alba_gpt/AlbaGPT_main/contents/response/SERVING_response.json
@@ -1,63 +1,0 @@
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블에 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블에 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블로 신선한 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블에 신선한 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블에 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블에 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블에 신선한 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블에 신선한 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블에 신선한 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
